chore: split out hooks to separate modules
